usnic: Change tiny_mtu to a size_t (it's compared against an unsigned value)
Authored-by: Jeff Squyres <jsquyres@cisco.com>

cmr=v1.7.5:ticket=trac:4253

This commit was SVN r30836.

The following Trac tickets were found above:
  Ticket 4253 --> https://svn.open-mpi.org/trac/ompi/ticket/4253

+++ ompi/mca/btl/usnic/btl_usnic_module.h
@@ -129,7 +129,7 @@ typedef struct ompi_btl_usnic_module_t {
      * segment is slightly less than what can be held in frag segment due
      * to fragment reassembly info.
      */
-    int tiny_mtu;
+    size_t tiny_mtu;
     size_t max_frag_payload;    /* most that fits in a frag segment */
     size_t max_chunk_payload;   /* most that can fit in chunk segment */
     size_t max_tiny_payload;    /* threshold for using inline send */
